requirements/local.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --resolver=backtracking requirements/local.in
#
alabaster==0.7.13
    # via sphinx
anyio==3.6.2
    # via
    #   httpcore
    #   watchfiles
appnope==0.1.3
    # via ipython
argon2-cffi==21.3.0
    # via -r requirements/base.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.6.0
    # via django
astroid==2.15.2
    # via pylint
asttokens==2.2.1
    # via stack-data
async-timeout==4.0.2
    # via redis
attrs==23.1.0
    # via -r requirements/base.in
babel==2.12.1
    # via sphinx
backcall==0.2.0
    # via ipython
black==23.3.0
    # via -r requirements/local.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
cfgv==3.3.1
    # via pre-commit
charset-normalizer==3.1.0
    # via requests
click==8.1.3
    # via
    #   black
    #   uvicorn
clients==1.4
    # via -r requirements/base.in
colorama==0.4.6
    # via sphinx-autobuild
coverage==7.2.3
    # via
    #   -r requirements/local.in
    #   django-coverage-plugin
crispy-bootstrap5==0.7
    # via -r requirements/base.in
cryptography==40.0.2
    # via pyjwt
decorator==5.1.1
    # via
    #   ipdb
    #   ipython
defusedxml==0.7.1
    # via python3-openid
dill==0.3.6
    # via pylint
distlib==0.3.6
    # via virtualenv
django==4.2
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap5
    #   django-allauth
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-model-utils
    #   django-redis
    #   django-stubs
    #   django-stubs-ext
django-allauth==0.54.0
    # via -r requirements/base.in
django-coverage-plugin==3.0.0
    # via -r requirements/local.in
django-crispy-forms==2.0
    # via
    #   -r requirements/base.in
    #   crispy-bootstrap5
django-debug-toolbar==4.0.0
    # via
    #   -r requirements/local.in
    #   django-requests-debug-toolbar
django-environ==0.10.0
    # via -r requirements/base.in
django-extensions==3.2.1
    # via -r requirements/local.in
django-model-utils==4.3.1
    # via -r requirements/base.in
django-redis==5.2.0
    # via -r requirements/base.in
django-requests-debug-toolbar==0.0.3
    # via -r requirements/local.in
django-stubs==1.16.0
    # via -r requirements/local.in
django-stubs-ext==0.8.0
    # via django-stubs
docutils==0.19
    # via sphinx
executing==1.2.0
    # via stack-data
factory-boy==3.2.1
    # via -r requirements/local.in
faker==18.4.0
    # via factory-boy
filelock==3.11.0
    # via virtualenv
flake8==6.0.0
    # via
    #   -r requirements/local.in
    #   flake8-isort
flake8-isort==6.0.0
    # via -r requirements/local.in
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
hiredis==2.2.2
    # via -r requirements/base.in
httpcore==0.17.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.24.0
    # via clients
identify==2.5.22
    # via pre-commit
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
imagesize==1.4.1
    # via sphinx
iniconfig==2.0.0
    # via pytest
ipdb==0.13.13
    # via -r requirements/local.in
ipython==8.12.0
    # via ipdb
isort==5.12.0
    # via
    #   flake8-isort
    #   pylint
jedi==0.18.2
    # via ipython
jinja2==3.1.2
    # via sphinx
lazy-object-proxy==1.9.0
    # via astroid
livereload==2.6.3
    # via sphinx-autobuild
markupsafe==2.1.2
    # via
    #   jinja2
    #   werkzeug
matplotlib-inline==0.1.6
    # via ipython
mccabe==0.7.0
    # via
    #   flake8
    #   pylint
mypy==1.2.0
    # via
    #   -r requirements/local.in
    #   django-stubs
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.7.0
    # via pre-commit
oauthlib==3.2.2
    # via requests-oauthlib
packaging==23.1
    # via
    #   black
    #   pytest
    #   pytest-sugar
    #   sphinx
parso==0.8.3
    # via jedi
pathspec==0.11.1
    # via black
petl==1.7.12
    # via -r requirements/base.in
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pillow==9.5.0
    # via -r requirements/base.in
platformdirs==3.2.0
    # via
    #   black
    #   pylint
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==3.2.2
    # via -r requirements/local.in
prompt-toolkit==3.0.38
    # via ipython
psycopg2==2.9.6
    # via -r requirements/local.in
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.2
    # via stack-data
pycodestyle==2.10.0
    # via flake8
pycparser==2.21
    # via cffi
pyflakes==3.0.1
    # via flake8
pygments==2.15.0
    # via
    #   ipython
    #   sphinx
pyjwt[crypto]==2.6.0
    # via django-allauth
pylint==2.17.2
    # via
    #   pylint-django
    #   pylint-plugin-utils
pylint-django==2.5.3
    # via -r requirements/local.in
pylint-plugin-utils==0.7
    # via pylint-django
pympler==1.0.1
    # via -r requirements/local.in
pytest==7.3.1
    # via
    #   -r requirements/local.in
    #   pytest-django
    #   pytest-sugar
pytest-django==4.5.2
    # via -r requirements/local.in
pytest-sugar==0.9.7
    # via -r requirements/local.in
python-dateutil==2.8.2
    # via faker
python-dotenv==1.0.0
    # via uvicorn
python-slugify==8.0.1
    # via -r requirements/base.in
python3-openid==3.2.0
    # via django-allauth
pytz==2023.3
    # via -r requirements/base.in
pyyaml==6.0
    # via
    #   pre-commit
    #   uvicorn
redis==4.5.4
    # via
    #   -r requirements/base.in
    #   django-redis
requests==2.28.2
    # via
    #   -r requirements/base.in
    #   django-allauth
    #   requests-oauthlib
    #   sphinx
requests-oauthlib==1.3.1
    # via django-allauth
six==1.16.0
    # via
    #   asttokens
    #   livereload
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
snowballstemmer==2.2.0
    # via sphinx
sphinx==6.1.3
    # via
    #   -r requirements/local.in
    #   sphinx-autobuild
sphinx-autobuild==2021.3.14
    # via -r requirements/local.in
sphinxcontrib-applehelp==1.0.4
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.1
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.2
    # via ipython
termcolor==2.2.0
    # via pytest-sugar
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via django-stubs
tomlkit==0.11.7
    # via pylint
tornado==6.2
    # via livereload
traitlets==5.9.0
    # via
    #   ipython
    #   matplotlib-inline
types-pytz==2023.3.0.0
    # via django-stubs
types-pyyaml==6.0.12.9
    # via django-stubs
typing-extensions==4.5.0
    # via
    #   django-stubs
    #   django-stubs-ext
    #   mypy
urllib3==1.26.15
    # via requests
uvicorn[standard]==0.21.1
    # via -r requirements/base.in
uvloop==0.17.0
    # via uvicorn
virtualenv==20.21.0
    # via pre-commit
watchdog==3.0.0
    # via werkzeug
watchfiles==0.19.0
    # via
    #   -r requirements/local.in
    #   uvicorn
wcwidth==0.2.6
    # via prompt-toolkit
websockets==11.0.1
    # via uvicorn
werkzeug[watchdog]==2.2.3
    # via -r requirements/local.in
wrapt==1.15.0
    # via astroid

# The following packages are considered to be unsafe in a requirements file:
# setuptools
